+++ src/interfaces/bill.interface.ts
