+++ src/models/budget.model.ts
@@ -1,5 +1,5 @@
 import { model, Schema, Document } from 'mongoose';
-import {BudgetManager} from '../interfaces/budget';
+import {BudgetManager} from '../interfaces/budget.interface';
 import {billSchema} from './bill.model';
 
 const budgetSchema: Schema = new Schema({
